+++ CHANGELOG.md
@@ -12,7 +12,7 @@ Version numbers are based on [Semantic Versioning](https://semver.org).
 
 ### Updated
 
-- Build documentation only for committed files in pre-commit hook
+- Do not build documentation on pre-commit as it produces too many failures
 - Allow multiline texts as localizations in `l10n` XML files
 - Remove `DataUtility` class as data strings are provided by [`LocalizationUtility`](src/classes/Utility/LocalizationUtility.php) class
 - SlackController: Allow adding more data when building message attachment footer