+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Version numbers are based on [Semantic Versioning](https://semver.org).
 - Allow multiline texts as localizations in `l10n` XML files
 - Remove `DataUtility` class as data strings are provided by [`LocalizationUtility`](src/classes/Utility/LocalizationUtility.php) class
 - SlackController: Allow adding more data when building message attachment footer
+- SlackController: Allow definition of custom configuration when building bot message
 
 ### Fixed